+++ test/service.js
@@ -271,21 +271,4 @@ export class Service {
 			response: req.scope || "authentication succeeded!",
 		};
 	}
-
-	async postMultipleBodyMimeTypes(req) {
-		if (req.body.str1 !== "string" && req.body.int1 !== 2) {
-			throw new Error(
-				"req.body.str1 is not a string or req.body.int1 is not 2",
-			);
-		}
-		return "";
-	}
-	async getMultipleResponseMimeTypes(req, reply) {
-		if (req.query.responseType === "application/json") {
-			reply.type("application/json");
-			return { str1: "test data" };
-		}
-		reply.type("text/json");
-		return { int1: 2 };
-	}
 }
